internal/v1: recognize and ignore old client caveat
Livepatch-auth clients are widely deployed but do not
correctly implement later versions of the interaction protocol.
To ensure smooth upgrade, this change causes the currently
deployed version of the identity server to recognize but ignore
the special caveat (<is-authenticated-user) that will cause
the later Candid version to use the old interaction protocol.